docs-translator/cloud-run-app/requirements.txt
Flask==3.0.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.111.0
google-generativeai==0.7.2
gunicorn==21.2.0
requests==2.31.0
beautifulsoup4==4.12.2
markdownify==0.11.6
